feat(org-id): add ABI selector for wt contracts
