+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "@windingtree/off-chain-adapter-http": "^3.0.1",
     "@windingtree/off-chain-adapter-swarm": "^4.0.1",
+    "@windingtree/wt-contracts": "^0.7.0",
     "@windingtree/wt-js-libs": "^0.13.0",
     "@windingtree/wt-ui": "^1.5.1",
     "@windingtree/wt-ui-react": "^0.14.0",
@@ -33,6 +34,7 @@
     "react-hot-loader": "^4.11.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "^2.1.8",
+    "react-select": "^3.0.4",
     "web3-eth": "^1.0.0-beta.35",
     "web3-utils": "^1.0.0-beta.35"
   },
