+++ src/assets/styles.scss
@@ -19,17 +19,3 @@ select.form-control {
     padding: 1em;
   }
 }
-
-.caller {
-  margin-bottom: 150px;
-
-  .alert {
-    position: fixed;
-    top: 10px;
-    left: 10px;
-  }
-
-  .card {
-    max-width: 600px;
-  }
-}
